% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEFAULT_EMISS_DF}
\alias{DEFAULT_EMISS_DF}
\title{Hector inputs and emissions that remain constant regardless of GCAM run}
\format{
\subsection{\code{DEFAULT_EMISS_DF}}{

A data frame with 10,008 rows and 4 columns:
\describe{
\item{year}{year}
\item{variable}{hector input variable name}
\item{value}{value}
\item{units}{variable units}
}
}
}
\source{
\url{https://github.com/JGCRI/gcam-core/blob/master/input/climate/default_emissions.csv}
}
\usage{
DEFAULT_EMISS_DF
}
\description{
These input variables are set to some default value that does not
vary with GCAM run, these variables are not modeled by GCAM V8.2.
}
\keyword{datasets}
