% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMISS_MAP_DF}
\alias{EMISS_MAP_DF}
\title{Data frame of the GCAM variable to hector input mapping}
\format{
\subsection{\code{EMISS_MAP_DF}}{

A data frame with 10,008 rows and 4 columns:
\describe{
\item{ghg}{GCAM ghg name}
\item{agg.gas}{GCAM label}
\item{unit.conv}{conversion factor to be applied to the GCAM results}
\item{hector.name}{hector variable name}
\item{hector.units}{hector variable units}
}
}
}
\usage{
EMISS_MAP_DF
}
\description{
Data frame of the GCAM variable to hector input mapping
}
\keyword{datasets}
