% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgcam.R
\name{get_all_queries}
\alias{get_all_queries}
\title{Process a GCAM xml output data base}
\usage{
get_all_queries(db_dir, db_name, query_file = NULL, prj_file = NULL)
}
\arguments{
\item{db_dir}{path to the directory that contains the GCAM xml db to process see \code{\link[rgcam]{localDBConn}}.}

\item{db_name}{str name of xml db to process see see \code{\link[rgcam]{localDBConn}}.}

\item{query_file}{XML query file to pass to the GCAM Model Interface. If NULL, use a default query file containing hector relevant queries.}

\item{prj_file}{R project data file to add extracted results to. Can be name of a project data file. The file will be created if it doesn't already exist. Default set to NULL will write the project data out to a temporary directory.}
}
\value{
the path to the rgcam project file.
}
\description{
Get all of the data from a GCAM xml output data base by running all
of the climate and emission queries.
}
