% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emiss.R
\name{get_hector_inputs}
\alias{get_hector_inputs}
\title{Prepare hector inputs from a GCAM xml output db}
\usage{
get_hector_inputs(
  db_dir,
  db_name,
  query_file = NULL,
  prj_file = NULL,
  gcam_emiss_file = NULL,
  gcam_default_file = NULL
)
}
\arguments{
\item{db_dir}{str directory where the GCAM database to be processed lives}

\item{db_name}{str name of the GCAM XML db}

\item{query_file}{str to the query file to run, default is set to internal data}

\item{prj_file}{str name where to save the rgcam project data at, if NULL will save to a temporary location}

\item{gcam_emiss_file}{str path to the hector gcam emissions csv table, default is set to internal data}

\item{gcam_default_file}{str path to the hector default emissions csv table, default is set to internal data}
}
\value{
data frame of Hector inputs
}
\description{
Prepare hector inputs from a GCAM xml output db
}
\examples{
\donotrun{

# Get the inputs for Hector using GCAM emissions during the
future period.
inputs <- get_hector_inputs(db_dir = "gcam_output",
                           db_name = "database_basexdb",
                           prj_file = "gcam_db.dat")
head(inputs)

# The data frame returned can be used be used in
}
}
