% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PREGCAM_EMISS_DF}
\alias{PREGCAM_EMISS_DF}
\title{Hector inputs and emissions used before GCAM}
\format{
\subsection{\code{PREGCAM_EMISS_DF}}{

A data frame with 6,003 rows and 4 columns:
\describe{
\item{year}{year}
\item{variable}{hector input variable name}
\item{value}{value}
\item{units}{variable units}
}
}
}
\source{
\url{https://github.com/JGCRI/gcam-core/blob/master/input/climate/gcam_emissions.csv}
}
\usage{
PREGCAM_EMISS_DF
}
\description{
The inputs that drive Hector from 1745 until the GCAM run begins
and future emissions are passed directly from GCAM to Hector.
}
\keyword{datasets}
