% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgcam.R
\name{fetch_GCAM_vs_hector}
\alias{fetch_GCAM_vs_hector}
\title{Get hector variables from the GCAM results}
\usage{
fetch_GCAM_vs_hector(prj_file, vars = NULL)
}
\arguments{
\item{prj_file}{path to the directory that contains the GCAM xml db created by \code{\link{get_all_queries}}.}

\item{vars}{hector variables to save, if NULL will return the CO2 concentrations, gmst, and a handful of radiative forcing results.}
}
\value{
data frame of the GCAM-hector results
}
\description{
Get hector variables from the GCAM results, to be used in a comparison
between stand alone hector and the GCAM-hector results.
}
